refactor: delete sequential-executor.ts, migrate tests to unified-executor
Production code already uses executeUnified (runner-execution.ts). This removes
the old sequential-executor.ts and migrates all 5 referencing test files:

- test/unit/execution/sequential-executor.test.ts → test/unit/interaction/triggers.test.ts
  (tests were actually for interaction/triggers, not the executor)
- sequential-executor-rl002.test.ts → unified-executor-rl002.test.ts
  (executeSequential → executeUnified; add missing failedTestFiles field to mocks)
- sequential-executor-rl007.test.ts → unified-executor-rl007.test.ts
  (source inspection tests now read unified-executor.ts; strip comments before
   checking finally blocks since unified-executor documents stopHeartbeat absence)
- integration/interaction/interaction-chain-pipeline.test.ts: update type import
- integration/execution/deferred-review-integration.test.ts: executeSequential → executeUnified
- integration/execution/parallel-batch.test.ts: remove dead executeSequential fallback in AC-18

All 1188 tests pass; typecheck and lint clean.
