+++ test/unit/interaction/triggers.test.ts
@@ -1,6 +1,6 @@
 /**
- * Unit tests for cost trigger wiring in sequential-executor.ts (TC-001)
- * and US-004: agentGetFn passthrough to preRunCtx.
+ * Unit tests for cost/merge trigger wiring (TC-001)
+ * and US-004: agentGetFn passthrough to preRunCtx in unified-executor.
  *
  * Covers: checkCostExceeded abort/skip/continue, checkCostWarning at 80%/100%
  * threshold, and isTriggerEnabled guard (no interaction plugin = today behavior).
@@ -118,12 +118,12 @@ describe("checkCostExceeded — abort response exits with cost-limit", () => {
 // US-004: agentGetFn passed to preRunCtx (structural verification)
 // ─────────────────────────────────────────────────────────────────────────────
 
-describe("US-004: sequential-executor passes agentGetFn to preRunCtx", () => {
-  test("sequential-executor.ts source includes agentGetFn: ctx.agentGetFn in preRunCtx", async () => {
+describe("US-004: unified-executor passes agentGetFn to preRunCtx", () => {
+  test("unified-executor.ts source includes agentGetFn: ctx.agentGetFn in preRunCtx", async () => {
     // Structural test: verify the source file wires agentGetFn into preRunCtx.
     // This prevents regressions where agentGetFn is removed from the context build.
     const source = await Bun.file(
-      new URL("../../../src/execution/sequential-executor.ts", import.meta.url).pathname,
+      new URL("../../../src/execution/unified-executor.ts", import.meta.url).pathname,
     ).text();
     expect(source).toContain("agentGetFn: ctx.agentGetFn");
   });
